+++ common/src/main/java/com/wynntils/core/features/FeatureManager.java
@@ -80,6 +80,7 @@ import com.wynntils.features.overlays.AuraTimerOverlayFeature;
 import com.wynntils.features.overlays.CombatExperienceOverlayFeature;
 import com.wynntils.features.overlays.ContentTrackerOverlayFeature;
 import com.wynntils.features.overlays.CustomBarsOverlayFeature;
+import com.wynntils.features.overlays.CustomPlayerListFeature;
 import com.wynntils.features.overlays.GameBarsOverlayFeature;
 import com.wynntils.features.overlays.GameNotificationOverlayFeature;
 import com.wynntils.features.overlays.GuildAttackTimerOverlayFeature;
@@ -119,7 +120,6 @@ import com.wynntils.features.ui.ContainerScrollFeature;
 import com.wynntils.features.ui.CosmeticsPreviewFeature;
 import com.wynntils.features.ui.CustomCharacterSelectionScreenFeature;
 import com.wynntils.features.ui.CustomLoadingScreenFeature;
-import com.wynntils.features.ui.CustomPlayerListFeature;
 import com.wynntils.features.ui.CustomSeaskipperScreenFeature;
 import com.wynntils.features.ui.LobbyUptimeFeature;
 import com.wynntils.features.ui.SoulPointTimerFeature;
@@ -241,6 +241,7 @@ public final class FeatureManager extends Manager {
         registerFeature(new CombatExperienceOverlayFeature());
         registerFeature(new ContentTrackerOverlayFeature());
         registerFeature(new CustomBarsOverlayFeature());
+        registerFeature(new CustomPlayerListFeature());
         registerFeature(new GameBarsOverlayFeature());
         registerFeature(new GameNotificationOverlayFeature());
         registerFeature(new GuildAttackTimerOverlayFeature());
@@ -292,7 +293,6 @@ public final class FeatureManager extends Manager {
         registerFeature(new CosmeticsPreviewFeature());
         registerFeature(new CustomCharacterSelectionScreenFeature());
         registerFeature(new CustomLoadingScreenFeature());
-        registerFeature(new CustomPlayerListFeature());
         registerFeature(new CustomSeaskipperScreenFeature());
         registerFeature(new LobbyUptimeFeature());
         registerFeature(new SoulPointTimerFeature());
